Universal Technical Copy for Directory MVP
Universal copy writing assets for Directory Service: Contains requests for technical copy needed on a screen-by-screen case derived from the task flows & interaction design documents. Includes a space to add that copy in both english and french. The idea is to have a central place to request, capture and share the necessary technical copy for use by developers, designers, product owners and translators.